+++ mne_connectivity/spectral/time.py
@@ -18,7 +18,7 @@ from mne.time_frequency import (
     tfr_array_morlet,
     tfr_array_multitaper,
 )
-from mne.utils import _check_option, _validate_type, logger, verbose
+from mne.utils import _check_option, _validate_type, logger, verbose, warn
 
 from ..base import EpochSpectralConnectivity, SpectralConnectivity
 from ..utils import _check_multivariate_indices, check_indices, fill_doc
@@ -44,9 +44,11 @@ def spectral_connectivity_time(
     average=False,
     indices=None,
     sfreq=None,
+    *,
     fmin=None,
     fmax=None,
-    fskip=0,
+    fskip=None,
+    fdecim=None,
     faverage=False,
     sm_times=0.0,
     sm_freqs=1,
@@ -131,8 +133,21 @@ def spectral_connectivity_time(
         The upper frequency of interest. Multiple bands are defined using a tuple, e.g.
         ``(13., 30.)`` for two band with 13 Hz and 30 Hz upper bounds. If ``None``, the
         highest frequency in ``freqs`` is used.
-    fskip : int
-        Omit every ``(fskip + 1)``-th frequency bin to decimate in frequency domain.
+    fskip : int | None
+        Omit every "(fskip + 1)-th" frequency bin to decimate in frequency domain. If
+        ``None`` (default) or 0, no frequency bins are skipped.
+
+        .. version-deprecated:: 0.8
+            ``fskip`` is deprecated and will be removed in 0.9. To reduce the number of
+            frequency bins, use ``fdecim`` instead, which offers more standard
+            decimation behaviour.
+    fdecim : int | None
+        Decimation factor in the frequency domain. Selects every Nth frequency bin from
+        the (time-)frequency decomposition (where N is the value of ``fdecim``). If
+        ``None`` (default) or 1, no decimation occurs. The default value will change to
+        1 in version 0.9.
+
+        .. versionadded:: 0.8
     faverage : bool
         Average connectivity scores for each frequency band. If ``True``, the output
         ``freqs`` will be an array of the average frequencies of each band.
@@ -186,8 +201,9 @@ def spectral_connectivity_time(
 
         .. versionadded:: 0.8
     decim : int
-        To reduce memory usage, decimation factor after time-frequency decomposition.
-        Returns ``tfr[…, ::decim]``.
+        Decimation factor in the time domain. Selects every Nth time bin from the
+        time-frequency decomposition (where N is the value of ``decim``). If 1
+        (default), no decimation occurs.
     n_jobs : int
         Number of connections to compute in parallel. Memory mapping must be activated.
         Please see the Notes section for details.
@@ -381,6 +397,19 @@ def spectral_connectivity_time(
     ----------
     .. footbibliography::
     """  # noqa: E501
+    if fskip is not None:
+        warn(
+            "The `fskip` parameter is deprecated and will be removed in 0.9. Use "
+            "`fdecim` instead.",
+            FutureWarning,
+        )
+        if fdecim is not None:
+            raise ValueError("`fskip` and `fdecim` cannot be used together.")
+    else:
+        fskip = 0
+    if fdecim is None:
+        fdecim = 1
+
     events = None
     event_id = None
     picks = None
@@ -494,6 +523,10 @@ def spectral_connectivity_time(
             "computing Granger causality on multiple frequency bands is not yet "
             "supported"
         )
+
+    _validate_type(fdecim, int, "fdecim", "int")
+    if fdecim < 1:
+        raise ValueError("`fdecim` must be >= 1")
 
     if any(this_method in _multivariate_methods for this_method in method):
         if not all(this_method in _multivariate_methods for this_method in method):
@@ -625,7 +658,7 @@ def spectral_connectivity_time(
         )
 
     # compute frequency mask based on specified min/max and decimation factor
-    freq_mask = _compute_freq_mask(freqs, fmin, fmax, fskip)
+    freq_mask = _compute_freq_mask(freqs, fmin, fmax, fskip, fdecim)
 
     # the frequency points where we compute connectivity
     freqs = freqs[freq_mask]